file structure refactor; rename sterp_response to calc_step_response
